requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mistune==3.0.0
python-frontmatter==1.0.0
PyYAML==6.0
Werkzeug==2.3.5
